only logged in user can update the listing

+++ app/Http/Controllers/ListingController.php
@@ -61,6 +61,11 @@ class ListingController extends Controller
     }
 
     public function update(Request $request, Listing $listing) {
+        // Make sure logged in user owns the listing
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -81,6 +86,10 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing) {
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action.');
+        }
+        
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing Deleted Successfully!');
